BeepController.cpp: Fixed possible casting bug: (signed) -> (signed long).
Although on my 32-bit Linux, ints and longs are the same length,
they might not be same length elsewhere,
especially on the Arduino the code is for.
So the (signed) cast to inappropriately reduce the resolution.
(signed long) avoids the resolution reduction.
"signed" in (signed long) is unnecessary,
since longs are signed.
I deliberately included "signed" in the (signed long) cast
to emphasis that it is signed.

+++ BeepController.cpp
@@ -70,7 +70,7 @@ void CBeepController::tick()
 		break;
 
 	case Beep_On:
-		if((signed)(millis() - m_ulTimer) >= 0L)	// Timer expired?
+		if((signed long)(millis() - m_ulTimer) >= 0L)	// Timer expired?
 		{
 			noTone(m_iBeepOutPin);				// Stop the tone
 			m_ulTimer = millis() + m_iOffTime;	// Set the off time
@@ -79,7 +79,7 @@ void CBeepController::tick()
 		break;
 
 	case Beep_Off:
-		if((signed)(millis() - m_ulTimer) >= 0L)	// Off timer expired?
+		if((signed long)(millis() - m_ulTimer) >= 0L)	// Off timer expired?
 		{
 			if(!m_bAlarm && (m_iRepeats > 0))	// Decrement the repeat counter?
 				--m_iRepeats;					// Not if there is an alarm
